Add files via upload
Decided that it would be best to keep the statistics tracking as a separate class, because it'll be easier to display data from the statistics and not interfere with the core game logic. I also made some changes to the Game and Dealer classes: I added statistics instances, made some new methods, and added some personality to the dealer lol. I'm making this a pull request, so you guys can look at it before it's added into the main branch.